InstanceTunnel: split off InstanceTunnelHandler (evaluation, no routing) as a base class
