+++ src/Mapbender/WmsBundle/Component/InstanceTunnelHandler.php
@@ -1,0 +1,89 @@
+<?php
+
+
+namespace Mapbender\WmsBundle\Component;
+
+use Mapbender\CoreBundle\Controller\ApplicationController;
+use Mapbender\CoreBundle\Entity\Source;
+use Mapbender\CoreBundle\Entity\SourceInstance;
+use Mapbender\CoreBundle\Utils\RequestUtil;
+use Symfony\Component\HttpFoundation\Request;
+
+
+/**
+ * Base version of InstanceTunnel that can process, but not generate tunnel requests (works without access to
+ * Router).
+ *
+ * @see ApplicationController::instanceTunnelAction()
+ *
+ * @package Mapbender\WmsBundle\Component
+ */
+class InstanceTunnelHandler
+{
+    /** @var SourceInstance */
+    protected $instance;
+
+    /** @var Source */
+    protected $source;
+
+    /**
+     * InstanceTunnel constructor.
+     * @param SourceInstance $instance
+     */
+    public function __construct(SourceInstance $instance)
+    {
+        $this->instance = $instance;
+        $this->source = $instance->getSource();
+    }
+
+    /**
+     * Gets the url on the wms service that satisfies the given $request (=Symfony Http Request object)
+     *
+     * @param Request $request
+     * @return string
+     */
+    public function getInternalUrl(Request $request)
+    {
+        $requestType = RequestUtil::getGetParamCaseInsensitive($request, 'request', null);
+
+        switch (strtolower($requestType)) {
+            case 'getmap':
+                return $this->source->getGetMap()->getHttpGet();
+            case 'getfeatureinfo':
+                return $this->source->getGetFeatureInfo()->getHttpGet();
+            case 'getlegendgraphic':
+                return $this->getInternalGetLegendGraphicUrl($request);
+            default:
+                return null;
+        }
+    }
+
+    /**
+     * Gets the url on the wms service that satisfies the given $request (=Symfony Http Request object)
+     *
+     * @param Request $request
+     * @return string
+     */
+    public function getInternalGetLegendGraphicUrl(Request $request)
+    {
+        $glgMode = $request->query->get('_glgmode', null);
+        $layerName = RequestUtil::getGetParamCaseInsensitive($request, 'layer', null);
+        if (!$layerName) {
+            $glgMode = null;
+            $layerSource = null;
+        } else {
+            $layerSource = WmsSourceEntityHandler::getLayerSourceByName($this->source, $layerName);
+            if (!$layerSource) {
+                $glgMode = null;
+            }
+        }
+        switch ($glgMode) {
+            default:
+                return $this->source->getGetLegendGraphic()->getHttpGet();
+            case 'styles':
+                return WmsInstanceLayerEntityHandler::getLegendUrlFromStyles($layerSource);
+            case 'GetLegendGraphic':
+                return WmsInstanceLayerEntityHandler::getLegendGraphicUrl($layerSource);
+        }
+    }
+}
