+++ src/Mapbender/WmsBundle/Component/WmsInstanceLayerEntityHandler.php
@@ -336,12 +336,11 @@ class WmsInstanceLayerEntityHandler extends SourceInstanceItemEntityHandler
     {
         $styleLegendUrl = $this->getLegendUrlFromStyles($entity->getSourceItem());
         if (WmsSourceEntityHandler::useTunnel($entity->getSourceInstance()->getSource())) {
-            $tunnel = new InstanceTunnel($this->entity->getSourceInstance());
             /** @var Router $router */
             $router = $this->container->get('router');
+            $tunnel = new InstanceTunnel($router, $entity->getSourceInstance());
         } else {
             $tunnel = null;
-            $router = null;
         }
         $layerName = $entity->getSourceItem()->getName();
         if ($styleLegendUrl) {
@@ -349,7 +348,7 @@ class WmsInstanceLayerEntityHandler extends SourceInstanceItemEntityHandler
                 // request via tunnel, see ApplicationController::instanceTunnelAction
                 // instruct the tunnel action that the legend url should be plucked from styles
                 $tunnelInputUrl = '?request=GetLegendGraphic&_glgmode=styles&layer=' . $layerName;
-                $publicLegendUrl = $tunnel->generatePublicUrl($router, $tunnelInputUrl);
+                $publicLegendUrl = $tunnel->generatePublicUrl($tunnelInputUrl);
             } else {
                 $publicLegendUrl = $styleLegendUrl;
             }
@@ -363,7 +362,7 @@ class WmsInstanceLayerEntityHandler extends SourceInstanceItemEntityHandler
                     // request via tunnel, see ApplicationController::instanceTunnelAction
                     // instruct the tunnel action that the legend url should be plucked from GetLegendGraphic
                     $tunnelInputUrl = '?request=GetLegendGraphic&_glgmode=GetLegendGraphic';
-                    $publicLegendUrl = $tunnel->generatePublicUrl($router, $tunnelInputUrl);
+                    $publicLegendUrl = $tunnel->generatePublicUrl($tunnelInputUrl);
                 } else {
                     $publicLegendUrl = $glgLegendUrl;
                 }
